Change type Value int to float fix bug no show detail forecast5day

+++ app/src/main/java/com/tea/weathernews/model/Main.java
@@ -2,7 +2,7 @@ package com.tea.weathernews.model;
 
 public class Main {
     private double temp;
-    private int pressure;
+    private float pressure;
     private int humidity;
     private double temp_min;
     private double temp_max;
@@ -18,11 +18,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
